+++ exi/projectExiConnector.h
@@ -18,8 +18,8 @@ extern struct dinEXIDocument dinDocEnc; /* The DIN document. For encoder. */
 extern struct dinEXIDocument dinDocDec; /* The DIN document. For decoder. */
 extern bitstream_t global_streamEnc; /* The byte stream descriptor. */
 extern bitstream_t global_streamDec; /* The byte stream descriptor. */
-extern size_t global_streamEncPos; /* The position in the stream. */
-extern size_t global_streamDecPos; /* The position in the stream. */
+extern uint32_t global_streamEncPos; /* The position in the stream. */
+extern uint32_t global_streamDecPos; /* The position in the stream. */
 extern int g_errn;
 
 #define SESSIONID_LEN 8
@@ -68,7 +68,7 @@ extern "C"
 {
 #endif
 void projectExiConnector_encode_DinExiDocument(void);
-  /* precondition: dinDocEnc structure is filled. Output: global_stream.data and global_stream.pos. */  
+  /* precondition: dinDocEnc structure is filled. Output: global_stream.data and global_stream.pos. */
 #if defined(__cplusplus)
 }
 #endif
